extended Board class with static functions for bit/square conversions

+++ src/damexchange/Protocol.h
@@ -5,8 +5,8 @@ namespace damexchange {
 // type traits
 template<typename Protocol> struct port                 { static const unsigned short value = 27531; };
 template<typename Protocol> struct terminator           { static const char           value =  '\0'; };
-template<typename Protocol> struct header_length        { static const size_t         value =     1; };
-template<typename Protocol> struct max_body_length      { static const size_t         value =   126; };
+template<typename Protocol> struct header_length        { static const int            value =     1; };
+template<typename Protocol> struct max_body_length      { static const int            value =   126; };
 
 struct protocol_V1 {};                  // current version of the DamExchange protocol
 typedef protocol_V1 protocol_tag;       // alias for the default protocol version
